requirements.txt
django
djangorestframework
django-filter
djangorestframework-simplejwt
django-cors-headers


